refactor: 당첨번호 분리
