+++ src/test/java/lotto/LottoTest.java
@@ -30,25 +30,4 @@ public class LottoTest {
 		assertThatIllegalArgumentException().isThrownBy(() -> new Lotto(Arrays.asList(1, 1, 2, 3, 4, 5)))
 			.withMessage("잘못된 번호 모음입니다.");
 	}
-
-	@Test
-	@DisplayName("당첨 테스트")
-	void prize() {
-		Lotto win = new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6));
-		Lotto six = new Lotto(Arrays.asList(6, 4, 2, 1, 5, 3));
-		Lotto five = new Lotto(Arrays.asList(6, 4, 2, 1, 5, 13));
-		Lotto four = new Lotto(Arrays.asList(6, 4, 2, 1, 15, 13));
-		Lotto three = new Lotto(Arrays.asList(6, 4, 2, 11, 15, 13));
-		Lotto two = new Lotto(Arrays.asList(6, 4, 12, 11, 15, 13));
-		Lotto one = new Lotto(Arrays.asList(6, 14, 12, 11, 15, 13));
-		Lotto zero = new Lotto(Arrays.asList(16, 14, 12, 11, 15, 13));
-
-		assertThat(six.result(win)).isEqualTo(Prize.SIX);
-		assertThat(five.result(win)).isEqualTo(Prize.FIVE);
-		assertThat(four.result(win)).isEqualTo(Prize.FOUR);
-		assertThat(three.result(win)).isEqualTo(Prize.THREE);
-		assertThat(two.result(win)).isEqualTo(Prize.NOTHING);
-		assertThat(one.result(win)).isEqualTo(Prize.NOTHING);
-		assertThat(zero.result(win)).isEqualTo(Prize.NOTHING);
-	}
 }
